Add v/oct mapping and 12-tone quantize
I don't understand the way CV input works exactly
but fmapping the input to 0v - 5v perfectly
aligns the CV output of my Keystep 37 with the
CV output. i.e. passthrough will relay the
correct note. Notes are divided by 1/12 and
rounded to get a quantization to the 12-tone
scale. Further work is needed to eliminate
notes and create other scales that input
can be quantized to.